Split Name::PropagateGenericClassifications tests out of name_test.rb (#4708)
Moves the 2 tests exercising Name::PropagateGenericClassifications
(can_propagate?, propagate_generic_classifications) into
test/models/name/propagate_generic_classifications_test.rb, matching
app/models/name/propagate_generic_classifications.rb.

require_relative'd from name_test.rb so running that file alone still
runs the full original suite.
